Remove manual hash validation
Prior to this change, updating the version of Prometheus or Alertmanager
meant changing the version number in several places in the installation
script as well as manually providing the hash for the tar files (see
https://github.com/18F/identity-idva-monitoring/pull/45/files). With
this change, users can update the version_num variable in one place at
the top of the script, and the URLs for the tar and the shasum file are
built by the script. This eliminates the possibility for human error and
typos and will always result in the version specified being installed.