Increase default memory to 256MB (#58)
Using 128MB is, by all accounts, a very small amount of memory for
prometheus to be using. Given that we have series in the thousands to
tens of thousands there's a small chance that we'll never see any issues
related to this, but as the number of series we ingest from our apps
grows, or the frequency increases, we will definitely start to hit the
memory threshold here. Increasing to 256 gives us a healthy amount of
growing room to support more data as well as the increase in memory
caused by our planned usage of the remote_write integration which
generally causes roughly 25% increase in utilization (see
https://prometheus.io/docs/practices/remote_write/#memory-usage).