Add external labels to support long-term storage
In order to support working with Cortex and to de-duplicate metric
streams, Prometheus needs to add labels to the data when communicating
with external systems. In the case of Cortex, it needs to know which
Prometheus instance sent the data. Following the guide at
https://cortexmetrics.io/docs/guides/ha-pair-handling/#client-side,
adding the "cluster" and "__replica__" labels should suffice. Note
however that only the "__replica__" label should be unique.