Remove "instance" label from alert labels (#123)
Removal of the "instance" label should solve the issue of multiple
alerts coming in for apps where multiple instances are running. IDVA
does not, nor should it ever, rely on the IP addresses found in the
"instance" label when deciding whether or not to send an alert. Dropping
this label when sending alerts to alertmanager allows us to receive
fewer alerts while preserving the data in Prometheus should teams need
to troubleshoot and identify problematic app instances.