Use static_configs for specific scrape jobs
In order to run the redis, elasticsearch, cf-metrics, and watchtower
apps in an HA configuration without over-burdening the monitoring system
and producing superfluous, duplicate data, we can use a static_config
scrape configuration. By using the static_config, prometheus will simply
hit the configured endpoint where it will then be load balanced to one
of the available app instances. This allows us to gather all of the data
from these exporters with the same frequency and level of detail, but
with less work on Prometheus's part and with no duplicate metric data
being ingested. It also has the secondary affect of allowing us to
horizontally scale out each of the mentioned services with zero affect
on the rest of the monitoring system.