improve monitoring availability (#156)
* Clarify hardware metrics data source

* Use static_configs for specific scrape jobs

In order to run the redis, elasticsearch, cf-metrics, and watchtower
apps in an HA configuration without over-burdening the monitoring system
and producing superfluous, duplicate data, we can use a static_config
scrape configuration. By using the static_config, prometheus will simply
hit the configured endpoint where it will then be load balanced to one
of the available app instances. This allows us to gather all of the data
from these exporters with the same frequency and level of detail, but
with less work on Prometheus's part and with no duplicate metric data
being ingested. It also has the secondary affect of allowing us to
horizontally scale out each of the mentioned services with zero affect
on the rest of the monitoring system.

* Increase replicas of Redis exporter

Due to the modifications in the prometheus scrape config to not scape by
individual instances, and instead use the app route to hit a single
instance, the redis exporter app can be scaled up with no issue.
Increasing the instances to > 1 now allows for high availability of the
app.

* Reduce memory allocation of Redis exporter

Over the last 14d, the Redis exporter has not consumed more than 35% of
it's allocated 64MB of memory. Reducing it by half saves on overall
system resources while still giving it a healthy amount of headroom in
case of spikes.

* Increase replicas of Elasticsearch exporter

Due to the modifications in the prometheus scrape config to not scape by
individual instances, and instead use the app route to hit a single
instance, the Elasticsearch exporter app can be scaled up with no issue.
Increasing the instances to > 1 now allows for high availability of the
app.

* Reduce memory allocation of Elasticsearch exporter

Over the last 14d, the Elasticsearch exporter has not consumed more than 35% of
it's allocated 64MB of memory. Reducing it by half saves on overall
system resources while still giving it a healthy amount of headroom in
case of spikes.

* Use static_configs for Grafana

* Increase Grafana instance replicas