Increase Cortex memory allocation
Over the last few week cortex memory utilization has hovered just over
the 80% threshold, causing a constant stream of alerts in our dev
environment. Increasing this to 512MB should circumvent the issue for
now. If the issue persists past this increase, we should look into what
is causing the memory to be consistently above thresholds.