Add multi-space support to network policy creation
In order to support apps in multiple spaces that need to communicate,
adding the -s option to the "cf add-network-policy" command allows the
script to designate the target space of the destination app. By
inspecting the target environment we can now create the desired network
policies independent of the space that happened to be targeted by the
app that triggered the policy creation.