watchtower use outbound proxy (#180)
* Improve .profile proxy initialization

The .profile script is being used to set the HTTP(S) proxy environment
variables for Watchtower, however several improvements needed to be
made.

1) Improve script output with echo commands such that the effect
   of the script is apparent within the cloud foundry application logs.

2) Use the "optional" operator "?" when accessing fields with jq that
   may not exist in the VCAP_SERVICES json object. This eliminates the
   cryptic "cannot iterate over null" messages when the app is not bound
   to the "outbound-proxy" user-provided-service.

3) Use jq's "-r" raw-output specifier when setting the "proxy_url"
   variable so that quotes are not included in the url-string. This was
   causing issues with go correctly parsing the URL from the
   environment.

* Add environment_name variable to network policy setup

* Add outbound-proxy to service bindings

* Move Watchtower installation outside of Cloud.gov

* Add multi-space support to network policy creation

In order to support apps in multiple spaces that need to communicate,
adding the -s option to the "cf add-network-policy" command allows the
script to designate the target space of the destination app. By
inspecting the target environment we can now create the desired network
policies independent of the space that happened to be targeted by the
app that triggered the policy creation.