Remove basic auth (#226)
The IDVA team has never made its Grafana instance publicly accessible,
and the use of basic auth was simply because it was there. The Grafana
application is already protected by the standard GSA authentication
mechanisms and requires MFA. We are removing basic auth as it
contributes little to the overall security posture of the service and
reduces the number of secrets required for the IVDA team to manage.